repr,sql-parser: reverse dependency edge
Reverse the dependency edge between repr and sql-parser so that
sql-parser depends on repr, rather than the other way around, as repr is
intended to be SQL agnostic.

Fix #1006.
