+++ src/repr/Cargo.toml
@@ -23,7 +23,6 @@ regex = "1.3.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1"
 serde_regex = "0.4.0"
-sql-parser = { path = "../sql-parser" }
 
 [dev-dependencies]
 criterion = "0.3"
