+++ src/repr/scalar/datetime.rs
@@ -7,14 +7,130 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+use std::fmt;
+use std::str::FromStr;
 use std::time::Duration;
 
 use failure::{bail, ensure, format_err};
 
 use super::Interval;
-use sql_parser::ast::DateTimeField;
 
 type Result<T> = std::result::Result<T, failure::Error>;
+
+#[derive(Debug, Clone, Copy, PartialOrd, Ord, PartialEq, Eq, Hash)]
+pub enum DateTimeField {
+    Year,
+    Month,
+    Day,
+    Hour,
+    Minute,
+    Second,
+}
+
+impl fmt::Display for DateTimeField {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.write_str(match self {
+            DateTimeField::Year => "YEAR",
+            DateTimeField::Month => "MONTH",
+            DateTimeField::Day => "DAY",
+            DateTimeField::Hour => "HOUR",
+            DateTimeField::Minute => "MINUTE",
+            DateTimeField::Second => "SECOND",
+        })
+    }
+}
+
+/// Iterate over `DateTimeField`s in descending significance
+impl IntoIterator for DateTimeField {
+    type Item = DateTimeField;
+    type IntoIter = DateTimeFieldIterator;
+    fn into_iter(self) -> DateTimeFieldIterator {
+        DateTimeFieldIterator(Some(self))
+    }
+}
+
+impl FromStr for DateTimeField {
+    type Err = String;
+
+    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
+        match s.to_uppercase().as_ref() {
+            "YEAR" | "YEARS" | "Y" => Ok(Self::Year),
+            "MONTH" | "MONTHS" | "MON" | "MONS" => Ok(Self::Month),
+            "DAY" | "DAYS" | "D" => Ok(Self::Day),
+            "HOUR" | "HOURS" | "H" => Ok(Self::Hour),
+            "MINUTE" | "MINUTES" | "M" => Ok(Self::Minute),
+            "SECOND" | "SECONDS" | "S" => Ok(Self::Second),
+            _ => Err(format!("invalid DateTimeField: {}", s)),
+        }
+    }
+}
+
+impl DateTimeField {
+    /// Iterate the DateTimeField to the next value.
+    /// # Panics
+    /// - When called on Second
+    pub fn next_smallest(self) -> Self {
+        self.into_iter()
+            .next()
+            .unwrap_or_else(|| panic!("Cannot get smaller DateTimeField than {}", self))
+    }
+    /// Iterate the DateTimeField to the prior value.
+    /// # Panics
+    /// - When called on Year
+    pub fn next_largest(self) -> Self {
+        self.into_iter()
+            .next_back()
+            .unwrap_or_else(|| panic!("Cannot get larger DateTimeField than {}", self))
+    }
+}
+
+/// An iterator over DateTimeFields
+///
+/// Always starts with the value smaller than the current one.
+///
+/// ```
+/// use repr::datetime::DateTimeField::*;
+/// let mut itr = Hour.into_iter();
+/// assert_eq!(itr.next(), Some(Minute));
+/// assert_eq!(itr.next(), Some(Second));
+/// assert_eq!(itr.next(), None);
+/// ```
+#[derive(Debug)]
+pub struct DateTimeFieldIterator(Option<DateTimeField>);
+
+/// Go through fields in descending significance order
+impl Iterator for DateTimeFieldIterator {
+    type Item = DateTimeField;
+    fn next(&mut self) -> Option<Self::Item> {
+        use DateTimeField::*;
+        self.0 = match self.0 {
+            Some(Year) => Some(Month),
+            Some(Month) => Some(Day),
+            Some(Day) => Some(Hour),
+            Some(Hour) => Some(Minute),
+            Some(Minute) => Some(Second),
+            Some(Second) => None,
+            None => None,
+        };
+        self.0.clone()
+    }
+}
+
+impl DoubleEndedIterator for DateTimeFieldIterator {
+    fn next_back(&mut self) -> Option<Self::Item> {
+        use DateTimeField::*;
+        self.0 = match self.0 {
+            Some(Year) => None,
+            Some(Month) => Some(Year),
+            Some(Day) => Some(Month),
+            Some(Hour) => Some(Day),
+            Some(Minute) => Some(Hour),
+            Some(Second) => Some(Minute),
+            None => None,
+        };
+        self.0.clone()
+    }
+}
 
 /// Tracks a unit and a fraction from a parsed time-like string, e.g. INTERVAL
 /// '1.2' DAYS.
@@ -953,8 +1069,6 @@ pub(crate) enum TimeStrToken {
 /// - Any sequence of numeric characters cannot be cast into an i64.
 /// - Any non-alpha numeric character cannot be cast into a TimeStrToken, e.g. `%`.
 pub(crate) fn tokenize_time_str(value: &str) -> Result<Vec<TimeStrToken>> {
-    use std::str::FromStr;
-
     let mut toks = vec![];
     let mut num_buf = String::with_capacity(4);
     let mut char_buf = String::with_capacity(7);
