+++ src/sql-parser/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 failure = "0.1.6"
 log = "0.4.5"
+repr = { path = "../repr" }
 
 [dev-dependencies]
 matches = "0.1"
